Create and test function to remove selected items from array
Created a function over the course of several days that
took an array and several option arguments to be removed from
said array. Through nested looping and learning about rest
parameters and redefining the "arguments" object, completed the
task. Was stuck for a while on making a copy of an array to
return while maintaining an unchanged array for the sake of the
function. This one stumped me for a while.