Code and test palindrome function
Created a function that tests whether a string is
a palindrome or not by converting the string into
an array of each character in the string, filtering
out punctuation and spacing, the rejoining the array
into a string. The string is then tested against itself
flipped to determine if it is a palindrome.